+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# ???
+## New Feature
+
+* Improve error messages
+
+## Internal
+
+* update combine to 4
+
 # 0.0.8
 ## Bug Fix